backend/pirsdjango/requirements.txt
django
djangorestframework
django-rest-knox
requests